[FEATURE] Se agrega nuevas capas
