--- src/main/java/com/conexionmysql/mysqljdb/aplication/controllers/ResourceNotFoundException.java
@@ -1,5 +1,0 @@
-package com.conexionmysql.mysqljdb.aplication.controllers;
-
-public class ResourceNotFoundException {
-
-}
