+++ src/main/java/com/conexionmysql/mysqljdb/application/ports/in/CreateReservationUseCase.java
@@ -1,0 +1,5 @@
+package com.conexionmysql.mysqljdb.application.ports.in;
+
+public class CreateReservationUseCase {
+  
+}
