+++ src/main/java/com/conexionmysql/mysqljdb/application/ports/in/CreateUserCommand.java
@@ -1,0 +1,14 @@
+package com.conexionmysql.mysqljdb.application.ports.in;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class CreateUserCommand {
+  private String name;
+  private Long userStateId;
+  private String email;
+  private String password;
+}
