+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/adapters/UserRepositoryAdapter.java
@@ -1,0 +1,5 @@
+package com.conexionmysql.mysqljdb.infrastructure.adapters;
+
+public class UserRepositoryAdapter {
+  
+}
