+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/controllers/UserController.java
@@ -1,0 +1,11 @@
+package com.conexionmysql.mysqljdb.infrastructure.controllers;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/users")
+public class UserController {
+
+  
+}
