+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/jpa/entities/UserState.java
@@ -1,11 +1,9 @@
-package com.conexionmysql.mysqljdb.domain.models;
+package com.conexionmysql.mysqljdb.infrastructure.jpa.entities;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -14,24 +12,14 @@ import lombok.Setter;
 
 
 @Entity
-@Table(name = "rooms")
+@Table(name = "users_state")
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
-public class Room {
-
+public class UserState {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long room_id;
-
-  @OneToOne
-  @JoinColumn(name = "type_room_id")
-  private RoomType type_room_id;
-
-  @OneToOne
-  @JoinColumn(name = "state_room_id")
-  private RoomState state_room_id;
-  private String name;
-  private String description;
+  private Long user_state_id; 
+  private String state;
 }
