+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/jpa/repositories/RoomStateRepository.java
@@ -1,8 +1,9 @@
-package com.conexionmysql.mysqljdb.domain.services;
+package com.conexionmysql.mysqljdb.infrastructure.jpa.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.conexionmysql.mysqljdb.domain.models.RoomState;
+
+import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.RoomState;
 
 @Repository
 public interface RoomStateRepository extends JpaRepository<RoomState, Long> {
