+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/jpa/repositories/RoomTypeRepository.java
@@ -1,8 +1,9 @@
-package com.conexionmysql.mysqljdb.domain.services;
+package com.conexionmysql.mysqljdb.infrastructure.jpa.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.conexionmysql.mysqljdb.domain.models.RoomType;
+
+import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.RoomType;
 
 @Repository
 public interface RoomTypeRepository extends JpaRepository<RoomType, Long> {
