+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/jpa/repositories/UserRepository.java
@@ -1,9 +1,9 @@
-package com.conexionmysql.mysqljdb.domain.services;
+package com.conexionmysql.mysqljdb.infrastructure.jpa.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.conexionmysql.mysqljdb.domain.models.User;
+import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
