+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/jpa/repositories/UserStateRepository.java
@@ -1,9 +1,9 @@
-package com.conexionmysql.mysqljdb.domain.services;
+package com.conexionmysql.mysqljdb.infrastructure.jpa.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.conexionmysql.mysqljdb.domain.models.UserState;
+import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.UserState;
 
 @Repository
 public interface UserStateRepository extends JpaRepository<UserState, Long> {
